#64 인증 필요 API 경로 분리 — logout, 회원탈퇴 /auth → /users 이동
