+++ src/main/java/com/guegue/duty_checker/user/controller/UserController.java
@@ -1,5 +1,6 @@
 package com.guegue.duty_checker.user.controller;
 
+import com.guegue.duty_checker.auth.service.AuthService;
 import com.guegue.duty_checker.user.dto.UpdateDeviceTokenReqDto;
 import com.guegue.duty_checker.user.service.UserService;
 import io.swagger.v3.oas.annotations.Operation;
@@ -21,6 +22,7 @@ import org.springframework.web.bind.annotation.*;
 public class UserController {
 
     private final UserService userService;
+    private final AuthService authService;
 
     @Operation(summary = "FCM 디바이스 토큰 업데이트", description = "푸시 알림 수신을 위한 FCM 디바이스 토큰을 등록하거나 갱신합니다.")
     @ApiResponses({
@@ -34,5 +36,28 @@ public class UserController {
             @Valid @RequestBody UpdateDeviceTokenReqDto reqDto) {
         userService.updateFcmToken(phone, reqDto.getFcmToken());
         return ResponseEntity.ok().build();
+    }
+
+    @Operation(summary = "로그아웃", description = "현재 사용자의 Refresh Token을 무효화하여 로그아웃 처리합니다.")
+    @ApiResponses({
+            @ApiResponse(responseCode = "200", description = "로그아웃 성공"),
+            @ApiResponse(responseCode = "401", description = "인증되지 않은 요청")
+    })
+    @PostMapping("/logout")
+    public ResponseEntity<Void> logout(@AuthenticationPrincipal String phone) {
+        authService.logout(phone);
+        return ResponseEntity.ok().build();
+    }
+
+    @Operation(summary = "회원탈퇴", description = "현재 사용자의 계정을 탈퇴 처리합니다. 연결된 Connection과 CheckIn 데이터가 삭제되며 복구할 수 없습니다.")
+    @ApiResponses({
+            @ApiResponse(responseCode = "204", description = "회원탈퇴 성공"),
+            @ApiResponse(responseCode = "401", description = "인증되지 않은 요청"),
+            @ApiResponse(responseCode = "404", description = "사용자를 찾을 수 없음")
+    })
+    @DeleteMapping("/me")
+    public ResponseEntity<Void> withdraw(@AuthenticationPrincipal String phone) {
+        authService.withdraw(phone);
+        return ResponseEntity.noContent().build();
     }
 }
